+++ src/lua/blight.rs
@@ -2,7 +2,7 @@ use super::{constants::*, ui_event::UiEvent};
 use crate::event::Event;
 use crate::{model::Line, PROJECT_NAME, VERSION};
 use log::debug;
-use rlua::{Result as LuaResult, UserData, UserDataMethods, Variadic};
+use rlua::{AnyUserData, Result as LuaResult, UserData, UserDataMethods, Variadic};
 use std::sync::mpsc::Sender;
 
 #[derive(Clone)]
@@ -44,27 +44,30 @@ impl Blight {
 
 impl UserData for Blight {
     fn add_methods<'lua, M: UserDataMethods<'lua, Self>>(methods: &mut M) {
-        methods.add_method_mut("output", |_, this, strings: Variadic<String>| {
+        methods.add_function("output", |ctx, strings: Variadic<String>| {
+            let this_aux = ctx.globals().get::<_, AnyUserData>("blight")?;
+            let mut this = this_aux.borrow_mut::<Blight>()?;
             this.output_lines.push(Line::from(strings.join(" ")));
             Ok(())
         });
-        methods.add_method("terminal_dimensions", |_, this, _: ()| {
+        methods.add_function("terminal_dimensions", |ctx, _: ()| {
+            let this_aux = ctx.globals().get::<_, AnyUserData>("blight")?;
+            let this = this_aux.borrow::<Blight>()?;
             Ok(this.screen_dimensions)
         });
-        methods.add_method(
-            "bind",
-            |ctx, _, (cmd, callback): (String, rlua::Function)| {
-                let bind_table: rlua::Table = ctx.globals().get(COMMAND_BINDING_TABLE)?;
-                bind_table.set(cmd.to_lowercase(), callback)?;
-                Ok(())
-            },
-        );
-        methods.add_method("unbind", |ctx, _, cmd: String| {
+        methods.add_function("bind", |ctx, (cmd, callback): (String, rlua::Function)| {
+            let bind_table: rlua::Table = ctx.globals().get(COMMAND_BINDING_TABLE)?;
+            bind_table.set(cmd.to_lowercase(), callback)?;
+            Ok(())
+        });
+        methods.add_function("unbind", |ctx, cmd: String| {
             let bind_table: rlua::Table = ctx.globals().get(COMMAND_BINDING_TABLE)?;
             bind_table.set(cmd, rlua::Nil)?;
             Ok(())
         });
-        methods.add_method_mut("ui", |_, this, cmd: String| {
+        methods.add_function("ui", |ctx, cmd: String| -> rlua::Result<()> {
+            let this_aux = ctx.globals().get::<_, AnyUserData>("blight")?;
+            let mut this = this_aux.borrow_mut::<Blight>()?;
             let event: UiEvent = UiEvent::from(cmd.as_str());
             if let UiEvent::Unknown(cmd) = event {
                 this.main_writer
@@ -75,24 +78,32 @@ impl UserData for Blight {
             }
             Ok(())
         });
-        methods.add_method("debug", |_, _, strings: Variadic<String>| {
+        methods.add_function("debug", |_, strings: Variadic<String>| {
             debug!("{}", strings.join(" "));
             Ok(())
         });
-        methods.add_method("is_core_mode", |_, this, ()| Ok(this.core_mode));
-        methods.add_method("status_height", |_, this, height: u16| {
+        methods.add_function("is_core_mode", |ctx, ()| {
+            let this_aux = ctx.globals().get::<_, AnyUserData>("blight")?;
+            let this = this_aux.borrow::<Blight>()?;
+            Ok(this.core_mode)
+        });
+        methods.add_function("status_height", |ctx, height: u16| {
+            let this_aux = ctx.globals().get::<_, AnyUserData>("blight")?;
+            let this = this_aux.borrow::<Blight>()?;
             this.main_writer
                 .send(Event::StatusAreaHeight(height))
                 .unwrap();
             Ok(())
         });
-        methods.add_method("status_line", |_, this, (index, line): (usize, String)| {
+        methods.add_function("status_line", |ctx, (index, line): (usize, String)| {
+            let this_aux = ctx.globals().get::<_, AnyUserData>("blight")?;
+            let this = this_aux.borrow::<Blight>()?;
             this.main_writer
                 .send(Event::StatusLine(index, line))
                 .unwrap();
             Ok(())
         });
-        methods.add_method("version", |_, _, _: ()| -> LuaResult<(&str, &str)> {
+        methods.add_function("version", |_, _: ()| -> LuaResult<(&str, &str)> {
             Ok((PROJECT_NAME, VERSION))
         });
     }